自定义分表算法

+++ sharding-table-springboot/src/main/java/com/sniper/shardingtablepringboot/algorithm/MyPreciseShardingAlgorithm.java
@@ -8,10 +8,10 @@ import java.util.Collection;
 /**
  * 自定义分片算法
  */
-public class MyPreciseShardingAlgorithm implements PreciseShardingAlgorithm<Long> {
+public class MyPreciseShardingAlgorithm implements PreciseShardingAlgorithm<Integer> {
 
 	@Override
-	public String doSharding(Collection<String> availableTargetNames, PreciseShardingValue<Long> shardingValue) {
+	public String doSharding(Collection<String> availableTargetNames, PreciseShardingValue<Integer> shardingValue) {
 		for (String tableName : availableTargetNames) {
 			if (tableName.endsWith(shardingValue.getValue() % 4 + "")) {
 				return tableName;
